requirements.txt
django==5.1.4
djangorestframework==3.15.2
drf_spectacular==0.28.0
openpyxl==3.1.5